pyproject.toml
[project]
name = "fy-4b-tools"
version = "0.1.0"
description = "A complete toolchain for fetching, processing, and tiling Fengyun-4B (FY-4B) satellite full-disk images into web-ready map tiles."
authors = [
    { name = "Lanseria", email = "zhangchao564265135@hotmail.com" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cartopy>=0.25.0",
    "gdal==3.11.1",
    "matplotlib>=3.10.6",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "scipy>=1.16.1",
    "tqdm>=4.67.1",
]


[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true
